requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
groq==0.9.0
python-dotenv==1.0.1
